Android: fix improper usage of $(local-intermediates-dir)
The macro $(local-intermediates-dir) should not be used in the recipes
since it depends on LOCAL_MODULE and LOCAL_MODULE_CLASS which are not
well-defined in the recipes. Use $(intermediates-dir-for) instead.

Besides, we should not remove the whole intermediates dir. Otherwise
the import_includes in that dir may be removed. It causes other errors.